(unifying) move NEM types into one file
